Update TextAdventure.java
In this commit (v1.2), the entry room is complete. There are new boolean conditions that were added to the program that change as the user interacts with the environment. For the most part, it is structured in the same way as previous versions:

Chain of if else if statements, boolean conditioned while loop (to get the user to enter a correct prompt), and new options.

A short "help" feature was added to let the user know they should be using short & simple commands to interact.

Additionally, the sleep() methods have been removed from the program completely. The program flows better without forcing pauses for the user. (11-5-17)